+++ src/pages/product/add-update.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+/* 
+商品添加/更新的路由组件
+*/
+export default class ProductAddUpdate extends Component {
+  render() {
+    return (
+      <div>
+        ProductAddUpdate
+      </div>
+    )
+  }
+}
