+++ src/pages/product/detail.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from 'react'
+
+/* 
+商品详情路由组件
+*/
+export default class ProductDetail extends Component {
+  render() {
+    return (
+      <div>
+        ProductDetail
+      </div>
+    )
+  }
+}
